+++ src/plugins/cloudinary-analytics/index.js
@@ -1,10 +1,14 @@
-import connectCloudinaryAnalytics from 'cloudinary-video-analytics';
+import { connectCloudinaryAnalytics } from 'cloudinary-video-analytics';
 import { PLAYER_EVENT } from '../../utils/consts';
 
 class CloudinaryAnalytics {
   constructor(player) {
     this.player = player;
     this.cloudinaryAnalytics = connectCloudinaryAnalytics(this.player.videoElement);
+    this.currentVideMetadata = {
+      cloudName: null,
+      publicId: null
+    };
   }
 
   getMetadata = () => ({
@@ -16,7 +20,13 @@ class CloudinaryAnalytics {
     this.player.on(PLAYER_EVENT.SOURCE_CHANGED, () => {
       const metadata = this.getMetadata();
       if (metadata.cloudName && metadata.publicId) {
-        this.cloudinaryAnalytics.startManuallyNewVideoTracking(metadata);
+        this.currentVideMetadata = metadata;
+        this.cloudinaryAnalytics.startManualTracking(metadata, {
+          videoPlayerType: 'cloudinary video player',
+          videoPlayerVersion: VERSION
+        });
+      } else if (this.currentVideMetadata.cloudName !== metadata.cloudName || this.currentVideMetadata.publicId !== metadata.publicId) {
+        this.cloudinaryAnalytics.stopManualTracking();
       }
     });
   }